moving pageConfig call to NetworkUtils
